backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
numpy
transformers
torch